attempt to split abi.go into 2 files
